fix: Update precompiles bytecode to get around the EVM checks
As precompiles are implemented inside the Runtime, they don't have a bytecode, and
their account code is empty by default. However in Solidity calling a function of a
contract often automatically adds a check that the contract bytecode is non-empty.
For that reason a dummy code (0x60006000fd) can be inserted at the precompile address
to pass that check.